Adjusted the business rules for registering, editing and deleting users

+++ clean-architecture-dotnet.Application/Services/Users/UserService.cs
@@ -9,12 +9,20 @@ namespace clean_architecture_dotnet.Application.Services.Users
 {
     public class UserService : IUserService
     {
-        private IUserRepository _userRepository;
+        private readonly IUserRepository _userRepository;
+        private readonly IUserAddressRepository _userAddressRepository;
+        private readonly IUserContactRepository _userContactRepository;
         private readonly IMapper _mapper;
 
-        public UserService(IMapper mapper, IUserRepository userRepository)
+        public UserService(
+            IUserRepository userRepository,
+            IUserAddressRepository userAddressRepository,
+            IUserContactRepository contactRepository,
+            IMapper mapper)
         {
             _userRepository = userRepository;
+            _userAddressRepository = userAddressRepository;
+            _userContactRepository = contactRepository;
             _mapper = mapper;
         }
 
@@ -63,11 +71,42 @@ namespace clean_architecture_dotnet.Application.Services.Users
         {
             try
             {
+                if (user is null)
+                    return Result<UserViewModel>.Fail("There is missing information to perform user change.", 404);
+
+                if (user.Address is null)
+                    return Result<UserViewModel>.Fail("There is missing information to perform user address change.", 404);
+
+                if (user.Contact is null)
+                    return Result<UserViewModel>.Fail("There is missing information to perform user address change.", 404);
+
+                var userEntity = await _userRepository.GetById(user.Id);
+                var addressEntity = await _userAddressRepository.GetById(user.Address.Id);
+                var contactEntity = await _userContactRepository.GetById(user.Contact.Id);
+
+                if (userEntity is null)
+                    return Result<UserViewModel>.Fail("User not found.", 404);
+
+                if (addressEntity is null)
+                    return Result<UserViewModel>.Fail("Address not found.", 404);
+
+                if (contactEntity is null)
+                    return Result<UserViewModel>.Fail("Contact not found.", 404);
+
                 var mapUser = _mapper.Map<User>(user);
+                var mapAddress = _mapper.Map<UserAddress>(user.Address);
+                var mapContact = _mapper.Map<UserContact>(user.Contact);
 
-                var result = await _userRepository.Put(mapUser);
+                var resultUser = await _userRepository.Put(mapUser);
+                var resultAddress = await _userAddressRepository.Put(mapAddress);
+                var resultContact = await _userContactRepository.Put(mapContact);
 
-                var mapUserModel = _mapper.Map<UserViewModel>(result);
+                var mapUserModel = _mapper.Map<UserViewModel>(resultUser);
+                var mapAddressModel = _mapper.Map<UserAddressViewModel>(resultAddress);
+                var mapContactModel = _mapper.Map<UserContactViewModel>(resultContact);
+
+                mapUserModel.Address = mapAddressModel;
+                mapUserModel.Contact = mapContactModel;
 
                 return Result<UserViewModel>.Ok(mapUserModel);
             }
@@ -82,12 +121,29 @@ namespace clean_architecture_dotnet.Application.Services.Users
         {
             try
             {
+                if(user is null)
+                    return Result<UserViewModel>.Fail("There is missing information to perform user creation.", 404);
+
+                if (user.Address is null)
+                    return Result<UserViewModel>.Fail("There is missing information to perform user address creation.", 404);
+
+                if (user.Contact is null)
+                    return Result<UserViewModel>.Fail("There is missing information to perform user address creation.", 404);
 
                 var mapUser = _mapper.Map<User>(user);
+                var mapAddress = _mapper.Map<UserAddress>(user.Address);
+                var mapContact = _mapper.Map<UserContact>(user.Contact);
 
-                var result = await _userRepository.Post(mapUser);
+                var resultUser = await _userRepository.Post(mapUser);
+                var resultAddress = await _userAddressRepository.Post(mapAddress);
+                var resultContact = await _userContactRepository.Post(mapContact);
 
-                var mapUserModel = _mapper.Map<UserViewModel>(result);
+                var mapUserModel = _mapper.Map<UserViewModel>(resultUser);
+                var mapAddressModel = _mapper.Map<UserAddressViewModel>(resultAddress);
+                var mapContactModel = _mapper.Map<UserContactViewModel>(resultContact);
+
+                mapUserModel.Address = mapAddressModel;
+                mapUserModel.Contact = mapContactModel;
 
                 return Result<UserViewModel>.Ok(mapUserModel);
             }
@@ -102,12 +158,37 @@ namespace clean_architecture_dotnet.Application.Services.Users
         {
             try
             {
+                if (user is null)
+                    return Result<UserViewModel>.Fail("There is missing information to delete the user.", 404);
+
+                if (user.Address is null)
+                    return Result<UserViewModel>.Fail("There is missing information to delete the user address.", 404);
+
+                if (user.Contact is null)
+                    return Result<UserViewModel>.Fail("There is missing information to delete the user contact.", 404);
+
+                var userEntity = await _userRepository.GetById(user.Id);
+                var addressEntity = await _userAddressRepository.GetById(user.Address.Id);
+                var contactEntity = await _userContactRepository.GetById(user.Contact.Id);
+
+                if (userEntity is null)
+                    return Result<UserViewModel>.Fail("User not found.", 404);
+
+                if (addressEntity is null)
+                    return Result<UserViewModel>.Fail("Address not found.", 404);
+
+                if (contactEntity is null)
+                    return Result<UserViewModel>.Fail("Contact not found.", 404);
 
                 var mapUser = _mapper.Map<User>(user);
+                var mapAddress = _mapper.Map<UserAddress>(user.Address);
+                var mapContact = _mapper.Map<UserContact>(user.Contact);
 
-                var result = await _userRepository.Delete(mapUser);
+                var resultUser = await _userRepository.Delete(mapUser);
+                var resultAddress = await _userAddressRepository.Delete(mapAddress);
+                var resultContact = await _userContactRepository.Delete(mapContact);
 
-                var mapUserModel = _mapper.Map<UserViewModel>(result);
+                var mapUserModel = _mapper.Map<UserViewModel>(resultUser);
 
                 return Result<UserViewModel>.Ok(mapUserModel);
             }
